requirements.txt
pandas==2.2.0  
influxdb-client==1.38.0  
python-dotenv==1.0.0  